Generalise strain search functionality
This generalises previous work (#187) to create a better framework for searching datasets for a particular strain (sample) name.

The "databases", which are a single JSON file per pathogen, are created by the `./scripts/collect-search-results.js` and uploaded to a S3 bucket rather than stored in this repo. This has a number of advantages: it separates out the logic for creating the databases from building this site which both reduces the build time and allows for future work to automatically update these databases (JSON files) when the datasets update. Currently these are hosted on the nextstrain-scratch bucket, and updates must be done manually. Please note that currently this script only searches across the nextstrain-data bucket; future improvement would be to add more pathogen-specific logic - for example searching the different nCoV datasets detailed in #185.

There is a simple nextstrain.org/search page which lists a number of per-pathogen search pages, which are dynamically defined by the data in `static-site/search_pages.json`. Each search page fetches the necessary data from a url defined in that JSON. Currently seasonal-flu and nCoV are included, but future additions will be simple to add.

Added, but commented out, a link to the nCoV search page from the splash page. This should be uncommented when we are ready to share this functionality which, for now, requires knowing that the /search URL exists.
