+++ build.sh
@@ -20,11 +20,8 @@ main() {
             build-static;;
         auspice)
             build-auspice;;
-        generate-data)
-            generate-data;;
         all)
             echo "Running the nextstrain.org build script"
-            generate-data # must come before build-static, as that depends on output from this
             build-static
             build-auspice
             echo "Build complete. Next step: \"npm run server\"";;
@@ -47,12 +44,6 @@ build-auspice() {
     cd auspice-client
     ../node_modules/.bin/auspice build --verbose --extend ./customisations/config.json
     cd ..
-}
-
-generate-data() {
-    echo "Running scripts to generate data at start time"
-    node scripts/collect-strains-sars-cov-2.js
-    node scripts/collect-exclusions-sars-cov-2.js
 }
 
 main "$@"
