--- scripts/collect-exclusions-sars-cov-2.js
@@ -1,42 +1,0 @@
-#!/usr/bin/env node
-
-const fetch = require("node-fetch");
-const fs = require('fs');
-
-(async () => {
-  console.log(`Fetching the SARS-CoV-2 exclude-list...`);
-  let exclude = await fetch("https://raw.githubusercontent.com/nextstrain/ncov/master/config/exclude.txt");
-  exclude = await exclude.text();
-  exclude = exclude.split("\n")
-    .reduce(
-      (state, currentLine) => {
-        /* The text file has no schema for associating comments explaining why a sequence is excluded
-        to the strains themselves. It's mostly obvious to humans, but hard to code.
-        The approach employed here is a heuristic based on what's been added so far (~800 lines).
-        A "reason" is a commented line with more than one word (so that we ignore a commented sample name)
-        and this reason is associated with the next block of text that's _not_ interrupted by a new line.
-        That is, a new line resets any assoication between a "reason" and subsequent sample names. */
-        if (currentLine === '') {
-          state.currentReason = '';
-          return state;
-        }
-        if (currentLine.startsWith('#')) {
-          const uncommented = currentLine.replace(/^#\s*/, '');
-          if ((uncommented.match(/\s+/g)||[]).length>0 && !state.currentReason) {
-            state.currentReason = uncommented;
-          }
-          // else it looks like a commented strain name, or a subsequent comment in a block -- do nothing!
-          return state;
-        }
-        if (!currentLine.match(/\s+/g)) {
-          state.excludeMap[currentLine] = state.currentReason || 'unknown reason';
-          return state;
-        }
-        console.log("Unparsed exclude.txt line: ", currentLine);
-        return state;
-      },
-      {excludeMap: {}, currentReason: ''}
-    );
-  if (!fs.existsSync("./data")) fs.mkdirSync("./data");
-  fs.writeFileSync("./data/ncov-excluded-strains.json", JSON.stringify(exclude.excludeMap, null, 0));
-})();
